Fix review time counting and max timer display
The timer displayed during review will now stop counting up when it
reaches the maximal answer time and will be highlighted red to indicate
that it has reached its limit (the desktop client does both).

This commit also changes the technique used to account for a card's true
review time after a pause and resume of the reviewer. The new technique
shifts back the *current* time by the elapsed interval instead of trying
to shift the original start time by the elapsed time.This prevents clock
changes from impacting the elapsed time.

This resolves the long-standing negative review time bug for which
there was a workaround in place, and thus removed in this commit.

+++ AnkiDroid/src/main/java/com/ichi2/libanki/Sched.java
@@ -1149,7 +1149,7 @@ public class Sched {
     }
 
 
-    private void log(long id, int usn, int ease, int ivl, int lastIvl, int factor, long timeTaken, int type) {
+    private void log(long id, int usn, int ease, int ivl, int lastIvl, int factor, int timeTaken, int type) {
         try {
             mCol.getDb().execute("INSERT INTO revlog VALUES (?,?,?,?,?,?,?,?,?)",
                     new Object[]{Utils.now() * 1000, id, usn, ease, ivl, lastIvl, factor, timeTaken, type});
